+++ packages/storybook/src/stories/Colors.tsx
@@ -1,4 +1,5 @@
 import { Fragment } from 'react'
+import { grey10, grey20 } from '@littlespoon/theme/lib/colors/secondary'
 
 interface Props {
   color: {
@@ -14,11 +15,13 @@ export default function Colors(props: Props) {
       {Object.entries(props.color).map(([colorGroup, colors], index) => (
         <Fragment key={colorGroup}>
           {index > 0 && <hr />}
+
           <section>
             <h1>{colorGroup}</h1>
+
             <ul style={{ listStyle: 'none' }}>
               {Object.entries(colors).map(([colorKey, color]) => (
-                <li key={colorKey}>
+                <li key={colorKey} style={{ display: 'flex' }}>
                   <div
                     style={{
                       backgroundColor: typeof color === 'function' ? color() : color,
@@ -28,7 +31,26 @@ export default function Colors(props: Props) {
                       width: 100,
                     }}
                   />
-                  <span>{colorKey}</span>
+
+                  <div>
+                    <p>
+                      <strong>{colorKey}</strong>
+                    </p>
+
+                    <pre
+                      style={{
+                        backgroundColor: grey10(),
+                        border: `.1rem solid ${grey20()}`,
+                        borderRadius: 4,
+                        display: 'inline',
+                        fontSize: '95%',
+                        padding: '.25rem .5rem',
+                        whiteSpace: 'nowrap',
+                      }}
+                    >
+                      <code>{typeof color === 'function' ? color() : color}</code>
+                    </pre>
+                  </div>
                 </li>
               ))}
             </ul>